Move RequestDetails in right directory.
