+++ coral/src/app/features/components/RequestDetailsModal.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import RequestDetailsModal from "src/app/features/approvals/components/RequestDetailsModal";
+import RequestDetailsModal from "src/app/features/components/RequestDetailsModal";
 
 const primaryActionText = "Approve";
 const primaryActionOnClick = jest.fn();
